[ios] Prompt for scoped permissions in client rather than standalone (#5117)
Follow up from https://github.com/expo/expo/pull/4839 as the underlying issue remained afterwards (see: https://twitter.com/elie2222/status/1156297725438304256)

I believe we meant to check that the appOwnership *is not* `expo` and in that case invoke the allow action, rather than do it when appOwnership *is* `expo`.

I tested this locally in Expo client by asking for a permission and forcing it to enter into the undetermined scope state so it would prompt.